Still More Changes, but i forgot what changes i made

+++ src/main/java/pages/DeleteQuestion.java
@@ -1,11 +1,9 @@
 package pages;
 
 import java.io.IOException;
-import javax.servlet.ServletException;
+import javax.servlet.*;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.*;
 
 import db.Database;
 
@@ -19,8 +17,12 @@ public class DeleteQuestion extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		String questionid = request.getParameter("id");
-		Database.deleteQuestion(questionid);
+		HttpSession session = request.getSession();
+		if (session.getAttribute("username") != null) {
+			String questionid = request.getParameter("id");
+			Database.deleteQuestion(questionid);
+		}
+
 		String referer = request.getHeader("referer");
 		int index = referer.indexOf("quiz.jsp");
 		response.sendRedirect(referer.substring(index));
